move wheel utils from test_utils to utils
